first working version for search & span search
