+++ backend/src/app/core/filters/abstract_column.py
@@ -1,9 +1,12 @@
 from enum import Enum, EnumMeta
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 from sqlalchemy.sql.base import ReadOnlyColumnCollection
 
 from app.core.filters.filtering_operators import FilterOperator, FilterValueType
+
+if TYPE_CHECKING:
+    from app.core.filters.search_builder import SearchBuilder
 
 
 class AbstractColumns(Enum, metaclass=EnumMeta):
@@ -26,4 +29,10 @@ class AbstractColumns(Enum, metaclass=EnumMeta):
         raise NotImplementedError
 
     def get_joins(self):
+        raise NotImplementedError
+
+    def add_subquery_filter_statements(self, query_builder: "SearchBuilder"):
+        raise NotImplementedError
+
+    def add_query_filter_statements(self, query_builder: "SearchBuilder"):
         raise NotImplementedError
