+++ backend/src/app/core/search/sdoc_search/sdoc_search_columns.py
@@ -1,6 +1,17 @@
+from sqlalchemy import String, cast, func
+from sqlalchemy.dialects.postgresql import ARRAY, array, array_agg
+
+from app.core.data.orm.annotation_document import AnnotationDocumentORM
+from app.core.data.orm.code import CodeORM
+from app.core.data.orm.document_tag import DocumentTagORM
 from app.core.data.orm.source_document import SourceDocumentORM
+from app.core.data.orm.span_annotation import SpanAnnotationORM
+from app.core.data.orm.span_text import SpanTextORM
+from app.core.data.orm.user import UserORM
+from app.core.db.sql_utils import aggregate_ids
 from app.core.filters.column_info import AbstractColumns
 from app.core.filters.filtering_operators import FilterOperator, FilterValueType
+from app.core.filters.search_builder import SearchBuilder
 
 
 class SearchColumns(str, AbstractColumns):
@@ -85,3 +96,70 @@ class SearchColumns(str, AbstractColumns):
                 return "Annotated by"
             case SearchColumns.SPAN_ANNOTATIONS:
                 return "Span annotations"
+
+    def add_subquery_filter_statements(self, query_builder: SearchBuilder):
+        match self:
+            case SearchColumns.DOCUMENT_TAG_ID_LIST:
+                query_builder._add_subquery_column(
+                    aggregate_ids(
+                        DocumentTagORM.id,
+                        label=SearchColumns.DOCUMENT_TAG_ID_LIST.value,
+                    )
+                )
+                query_builder._join_subquery(
+                    SourceDocumentORM.document_tags, isouter=True
+                )
+            case SearchColumns.CODE_ID_LIST:
+                query_builder._add_subquery_column(
+                    aggregate_ids(CodeORM.id, label=SearchColumns.CODE_ID_LIST.value)
+                )
+                query_builder._join_subquery(
+                    SourceDocumentORM.annotation_documents,
+                    isouter=True,
+                )
+                query_builder._join_subquery(
+                    SpanAnnotationORM.code,
+                    isouter=True,
+                )
+            case SearchColumns.USER_ID_LIST:
+                query_builder._add_subquery_column(
+                    aggregate_ids(UserORM.id, SearchColumns.USER_ID_LIST.value)
+                )
+                query_builder._join_subquery(
+                    SourceDocumentORM.annotation_documents,
+                    isouter=True,
+                )
+                query_builder._join_subquery(
+                    AnnotationDocumentORM.user,
+                    isouter=True,
+                )
+            case SearchColumns.SPAN_ANNOTATIONS:
+                query_builder._add_subquery_column(
+                    cast(
+                        array_agg(
+                            func.distinct(
+                                array([cast(CodeORM.id, String), SpanTextORM.text])
+                            ),
+                        ),
+                        ARRAY(String, dimensions=2),
+                    ).label(SearchColumns.SPAN_ANNOTATIONS.value)
+                )
+                query_builder._join_subquery(
+                    SourceDocumentORM.annotation_documents,
+                    isouter=True,
+                )
+                query_builder._join_subquery(
+                    AnnotationDocumentORM.span_annotations,
+                    isouter=True,
+                )
+                query_builder._join_subquery(
+                    SpanAnnotationORM.span_text,
+                    isouter=True,
+                )
+                query_builder._join_subquery(
+                    SpanAnnotationORM.code,
+                    isouter=True,
+                )
+
+    def add_query_filter_statements(self, query_builder: SearchBuilder):
+        pass
